requirements.txt
click==8.0.1
Flask==2.0.1
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
PyYAML==5.4.1
typing-extensions==3.10.0.2
Werkzeug==2.0.1
zipp==3.5.0
